requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.0
attrs==25.3.0
blinker==1.9.0
boto3==1.37.15
botocore==1.37.15
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
griffe==1.6.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
langchain-core==0.3.45
langgraph>=0.3.16
langsmith==0.3.17
logfire==3.9.0
logfire-api==3.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.5.1
msgpack==1.1.0
mypy-extensions==1.0.0
narwhals==1.31.0
numpy==2.2.4
openai==1.66.5
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.41
pydantic-ai-slim==0.0.41
pydantic-graph==0.0.41
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.46.1
streamlit==1.43.2
structlog==25.2.0
tenacity==9.0.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
